move acta photo handling to actions page
